Refactor Tag hierarchy
The List_Tag data is changed to use std::variant, specifically as
std::vector<payload_type>.

The simple tag data types Int_Tag, String_Tag, Float_Tag,
Byte_Array_Tag, etc. are eliminated.

In the
NBT hierarchy, they can only exist as children of either List_Tag or
Compound_Tag.  Furthermore, named tags can only be children of
Compound_Tag.

Therefore, their payloads are directly stored as payload_type in the
member data of either
List_Tag or Compound_Tag, and the names of the named tags are stored as
the keys in Compound_Tag.

Furthermore, the *_Array_Tag payloads are now stored as the std::vector variants
of the payload_type, rather than as std::span as was the case before.
This is for simplicity and could be changed in the future to minimize
heap allocation.

These changes resolve the refactoring project described by issue #2 on github.com.