Create std::optional NBT data getter method
I think this is the interface that I prefer to end up with.

Having to use exception handling or numerous explicit checks in order
to manage the presence of absence of optional entries in an NBT file
is verbose at best, and sometimes even quite a nuisance.

If users do not care to check the status of the optional return value,
they may still throw exceptions in much the same way, but really it's
their own choice at that point.

The main question I have remaining is: should the std::variant datatype
really encompass std::unique_ptr values?  [In order to avoid expensive
copies of Compound Tags and List Tags.]  If so, we must pass the
std::variant by reference.  Alternatively, the std::variant could be
used to hold only primitive values.  Or maybe one std::variant of
primitives and one of complex data types.

A fundamental problem is that we are trying to use std::variant to
hold unique_ptr, which has reference semantics.  Maybe the solution is
to either abandon std::variant (returning to a class hierarchy to encapsulate
nbt data, including primitive data types), or else to bite the bullet
and abandon unique_ptr.

Refer to:
https://www.foonathan.net/2022/05/recursive-variant-box/
https://stackoverflow.com/questions/72212173/how-to-define-recursive-stdvariant
https://www.reddit.com/r/cpp/comments/q42eyq/recursive_variant_a_recursive_variant_library_for/