Create specialized accessor Method
Ultimately, there should be two templated methods: one for value
semantics, one for reference semantics.  The first is for primitive
types, the second for complex types.  The references will be
const references derived from unique_ptr<T> held in the std::variant
data.  If the unique_ptr is null (which could happen e.g. due to
exception thrown during construction of the complex type) then
std::nullopt will be passed.  Similarly std::nullopt will be passed
for tags that are not found or are not of the type requested.