Complete accessor interface for Compound
get_ptr<T> is created to look up unique_ptr<T> in the variant map
and return a bare pointer if found.

The reason for the bare pointer
is that the accessor is provided only a view, not ownership in any way.
unique_ptr is uncopyable for a reason, and passing a const reference to
would have no advantage over a bare pointer.  A nullptr return
signifies that the name was either not found, or else was found but
not of type T.

get_opt<T> is renamed to get_value<T> to contrast with the meaning
of get_ptr<T>.  This is, it returns a value rather than a reference.

get_List, get_Compound, get_Byte_Array, etc. are all implemented
using get_ptr<T>.