Use string concatenation in to_string() methods
Instead of using ostringstream stream insertion to build the output
strings for the to_string() methods, we have switched to using
std::string::operator+= to build up the output strings.

Therefore we also remove the #include <stringstream> directive.

Additionally, we move the implementation of the to_string() functions
from nbtview.hpp to the nbtview.cpp file, and move their Helper
functions as well.

This is good because it make the nbtview.hpp interface file more
clear and lightweight as well.