Encapsulate deserialization into Deserializer
To improve separation of concerns, Tag, Compound, and List are no longer
responsible for deserialization.

Instead, we now follow the Builder pattern, so that subclasses of
Deserializer are normally responsible for creating and populating Tag
data structures from input data.

Client code can add to and modify Tag data through the provided
reference accessors.

In a future update, Stringification will be also be extracted from the
Tag, List, and Compound classes.