Add BinaryWriter
BinaryWriter writes primitive values in big endian form, suitable for
the NBT specification.  It also writes strings in NBT format (that is,
as a sequence of bytes preceded by a 2-byte length value).

In fact, write_string would be an appropriate place to throw an
exception for exceeding the maximum string length, but this is not yet done.