Make BinaryWriter a stateless, untemplated class
- Changes BinaryWriter to remove its member data, which had been templated
on an OutputIterator type.

- Changes BinaryWriter interface by adding the output stream as a
reference parameter to its methods.

This change removes unneeded dependencies by removing the need to
template the BinaryWriter class.  Its write() and write_vector()
functions remain templated for polymorphism of the output data.

The member functions are now all static, and no BinaryWriter instance is
needed for its use.

The tests are now changed to reflect the changes in the BinaryWriter interface.