Add Region data container
This implements just the header data (timestamps and offsets tables).

Access to chunk / NBT data will be added after header tests are written.

The design is open to allowing the rewriting of container data,
but this functionality is not needed for implementing my primary
objective, which is region file scraping from disk images.  So I am not
likely to add that feature anytime soon.