Add zlib header detection
Note that the header detection we're using is just checking for three
particular bytes.  These bytes are the actual header for any gzip
data (RFC 1952).  For zlib, it depends on the zlib format parameters
CINFO and FLEVEL.  Minecraft uses CINFO=7, FLEVEL=2, so the magic
numbers of 0x78, 0x9c should be the first two bytes of zlib encoded
data.