Add reading of chunk data.
Region is not concerned with the decoding of chunk data, so the data is
generally to be passed to nbt::read_binary().

Note that it is necessary for the Region's sector data to be populated,
which can be done via ReadOffsets.  However, client code may wish to set
the sector data through other means.  (E.g., merging different Region
files, etc.)

test_Chunks reads and decodes the chunks from the provided
test_data/r.0.0.mca file.

test_Chunks then extracts the xPos and zPos tags, which hold the
global chunk coordinates, and compares the values to the coordinates
local to the region, which corresponds to the offset's location
in the region header.