Feat: Require inflation from unsegmented data
The practical use case of representing compressed data as a vector of
spans of data (as in fragmented file extents) is not totally clear,
and it was significantly increasing the complexity and fragility of the
interface.

So, for now at least, we'll just require the data to be
contiguous.  When doing more advanced carving, this may
require more copying to make data contiguous, but if
that becomes a problem, we can add on a new interface to support
fragmentation of compressed data.