Feat: New data model
Compound and List are no longer type aliases for unique_ptr to STL
container class objects.

Instead Compound and List are classes that directly contain an STL map
and vector member respectively.

** This provides much greater flexibility for any future refactoring.
For example, if we want the Compound data to be held by a unique_ptr,
these changes will be restricted to the Compound class implementation
alone. **

Due to this change, necessary changes to the interface and
implementation of Tag and utility functions have been carried out as
well.